Allow history/countermove history to take full range
Don't saturate or truncate or anything, and don't decrease the full
range of the histories, let's use the full fat and see if this finally
lets countermove history work well.

Relies on some unfortunate coupling with the move sorting strata.

+++ src/history.h
@@ -14,7 +14,7 @@ void history_update(const Bitboard* board,
                     int8_t ply);
 const Move* history_get_killers(int8_t ply);
 Move history_get_countermove(const Bitboard* board);
-int16_t history_get_combined(const Bitboard* board, Move m);
+int32_t history_get_combined(const Bitboard* board, Move m);
 int16_t history_get_uncombined(Move m);
 
 #endif
